requirements.txt
numpy==1.26.4
opencv-contrib-python==4.8.1.78
pyrealsense2==2.54.2.5684
Pillow==10.2.0
reportlab==4.1.0
# matplotlib==3.8.0
# scipy==1.11.4
